Clearer language for MongoDB version
This language is more informative to make it clear that currently MongoDB 3.2 is NOT supported AND THUS neither are "joins".  Before if the reader didn't know the status of the current mongoDB version they'd have to break their reading to go look it up.  Especially with the word "typically" in the sentence made the old sentence particularly ambiguous as to the current state even though it is not referring to support but implementation choices.